requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
httpx==0.24.0
python-dotenv==1.0.0
openai==1.0.0
pydantic==1.10.12
numpy==1.26.0
SQLAlchemy==2.0.20
alembic==1.11.1
